# build/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"SigninParam": {
				"type": "string"
			},
			"SignupParam": {
				"properties": {
					"password": {
						"type": "string"
					},
					"login": {
						"type": "string"
					}
				},
				"type": "object"
			},
			"CreateBookParams": {
				"properties": {
					"totalPages": {
						"type": "number",
						"format": "double"
					},
					"author": {
						"type": "string"
					},
					"summary": {
						"type": "string"
					},
					"title": {
						"type": "string"
					}
				},
				"type": "object"
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "boilerplate-cleanarchi",
		"contact": {}
	},
	"paths": {
		"/users/signin": {
			"post": {
				"operationId": "Signin",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"accessToken": {
											"type": "string"
										}
									},
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "Invalid request params"
					},
					"404": {
						"description": "Not found"
					}
				},
				"summary": "Signin user",
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SigninParam"
							}
						}
					}
				}
			}
		},
		"/users/signup": {
			"post": {
				"operationId": "Signup",
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"accessToken": {
											"type": "string"
										}
									},
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "Invalid request params"
					},
					"409": {
						"description": "Already exists"
					}
				},
				"summary": "Signup user",
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SignupParam"
							}
						}
					}
				}
			}
		},
		"/books": {
			"get": {
				"operationId": "List",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"properties": {
											"totalPages": {
												"type": "number",
												"format": "double"
											},
											"author": {
												"type": "string"
											},
											"summary": {
												"type": "string"
											},
											"title": {
												"type": "string"
											},
											"id": {
												"type": "string"
											}
										},
										"type": "object"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"summary": "Get all books",
				"tags": [
					"Books"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"totalPages": {
											"type": "number",
											"format": "double"
										},
										"author": {
											"type": "string"
										},
										"summary": {
											"type": "string"
										},
										"title": {
											"type": "string"
										},
										"id": {
											"type": "string"
										}
									},
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "Invalid request params"
					}
				},
				"summary": "Create book",
				"tags": [
					"Books"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateBookParams"
							}
						}
					}
				}
			}
		},
		"/books/{id}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"totalPages": {
											"type": "number",
											"format": "double"
										},
										"author": {
											"type": "string"
										},
										"summary": {
											"type": "string"
										},
										"title": {
											"type": "string"
										},
										"id": {
											"type": "string"
										}
									},
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "Invalid book id format"
					},
					"404": {
						"description": "Not found"
					}
				},
				"summary": "Get a book by id",
				"tags": [
					"Books"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "The book's identifier",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"204": {
						"description": ""
					},
					"400": {
						"description": "Invalid book id format"
					},
					"404": {
						"description": "Not found"
					}
				},
				"summary": "Delete book",
				"tags": [
					"Books"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "The book's identifier",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "http://localhost:8001",
			"description": "Localhost"
		}
	]
}